cmd/tmsh-client: fix compilation and upgrade urfave/cli to the latest revision
For now the tmsh-client is not actively developed and remained at the
state of proof-of-concept. Some of the dependencies used will be open
source soon, until then I disabled everything that could prevent the
build so that our internal toolchain stops complaining about
unresolvable dependencies.

urfave/cli package has been upgraded to the latest revision in the
process in order to fix a vulnerability in one of its dependency
(gopkg.in/yaml.v2).